refactor(workout-plan): extract add-exercise cluster (WP3.4f)
